requirements.txt
asgiref==3.7.2
attrs==23.2.0
backports.zoneinfo==0.2.1
certifi==2024.2.2
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
colorama==0.4.6
Django==4.2.9
django-cors-headers==4.3.1
django-crispy-forms==2.1
django-filter==23.5
django-jazzmin==2.6.0
django-leaflet==0.29.0
djangorestframework==3.14.0
djangorestframework-gis==1.0
fiona==1.9.5
GDAL @ file:///C:/Users/LOUIS/Desktop/FINAL%20YEAR%20PROJECT%20WORK/Backend%20System/BackendAPI/GDAL-3.4.3-cp38-cp38-win_amd64.whl
geojson==3.1.0
geopandas==0.13.2
gunicorn==21.2.0
importlib-metadata==7.0.1
Markdown==3.5.2
numpy==1.24.4
packaging==23.2
pandas==2.0.3
psycopg2-binary==2.9.9
pyproj==3.5.0
python-dateutil==2.8.2
pytz==2024.1
shapely==2.0.3
six==1.16.0
sqlparse==0.4.4
typing-extensions==4.9.0
tzdata==2023.4
whitenoise==6.6.0
zipp==3.17.0
